+++ ports/stm32/boards/Passport/modules/flows/verify_backup_flow.py
@@ -4,9 +4,10 @@
 # verify_backup_flow.py - Verify a selected backup file.
 
 
-from constants import TOTAL_BACKUP_CODE_DIGITS
+from constants import NUM_BACKUP_PASSWORD_WORDS, TOTAL_BACKUP_CODE_DIGITS
 from flows import FilePickerFlow, Flow
-from pages import BackupCodePage, ErrorPage, InsertMicroSDPage, LongSuccessPage, SuccessPage
+from pages import (BackupCodePage, ErrorPage, InsertMicroSDPage, LongSuccessPage,
+                   PredictiveTextInputPage, QuestionPage, RecoveryModeChooserPage, SuccessPage)
 from utils import get_backup_code_as_password, get_backups_folder_path, spinner_task
 from tasks import verify_backup_task
 from errors import Error
@@ -18,6 +19,8 @@ class VerifyBackupFlow(Flow):
     def __init__(self):
         super().__init__(initial_state=self.choose_file, name='VerifyBackupFlow')
         self.backup_code = [None] * TOTAL_BACKUP_CODE_DIGITS
+        self.backup_password_words = []
+        self.backup_password_prefixes = []
         self.decryption_password = None
 
     async def choose_file(self):
@@ -25,20 +28,51 @@ class VerifyBackupFlow(Flow):
         result = await FilePickerFlow(initial_path=backups_path, suffix='.7z', show_folders=True).run()
         if result is None:
             # No file chosen, so go back to menu
-            self.clear_backup_code()
+            self.clear_credentials()
             self.set_result(False)
             return
 
         _filename, full_path, is_folder = result
         if not is_folder:
             self.backup_file_path = full_path
+            self.goto(self.select_recovery_mode)
+
+    async def select_recovery_mode(self):
+        from pages.recovery_mode_chooser_page import RecoveryMode
+
+        result = await RecoveryModeChooserPage(card_header={'title': 'Verify Backup'}).show()
+        if result is None:
+            self.clear_credentials()
+            self.back()
+        elif result == RecoveryMode.BACKUP_CODE_20_DIGITS:
             self.goto(self.enter_backup_code)
+        else:
+            self.goto(self.enter_backup_password)
+
+    async def enter_backup_password(self):
+        result = await PredictiveTextInputPage(
+            word_list='bytewords',
+            total_words=NUM_BACKUP_PASSWORD_WORDS,
+            initial_words=self.backup_password_words,
+            initial_prefixes=self.backup_password_prefixes).show()
+
+        (words, self.backup_password_prefixes, _) = result
+        if words is None:
+            cancel = await QuestionPage(text='Cancel password entry? All progress will be lost.').show()
+            if cancel:
+                self.clear_credentials()
+                self.back()
+        else:
+            self.backup_password_words = words
+            self.decryption_password = ' '.join(words)
+            self.goto(self.do_verify)
 
     async def enter_backup_code(self):
         result = await BackupCodePage(
             digits=self.backup_code,
             card_header={'title': 'Enter Backup Code'}).show()
         if result is None:
+            self.clear_credentials()
             self.back()
             return
 
@@ -52,18 +86,18 @@ class VerifyBackupFlow(Flow):
             verify_backup_task,
             args=[self.decryption_password, self.backup_file_path])
         if error is None:
-            self.clear_backup_code()
+            self.clear_credentials()
             page_class = SuccessPage if passport.IS_COLOR else LongSuccessPage
             await page_class(text='Backup decrypted successfully and passed its integrity check.').show()
             self.set_result(True)
         elif error is Error.MICROSD_CARD_MISSING:
             result = await InsertMicroSDPage().show()
             if not result:
-                self.clear_backup_code()
+                self.clear_credentials()
                 self.set_result(False)
         elif error is Error.INVALID_BACKUP_CODE:
             result = await ErrorPage(
-                text='Unable to decrypt backup. The Backup Code may be incorrect, '
+                text='Unable to decrypt backup. The Backup Code or password may be incorrect, '
                      'or the backup may be damaged.',
                 left_micron=microns.Back,
                 right_micron=microns.Retry).show()
@@ -71,25 +105,27 @@ class VerifyBackupFlow(Flow):
             if result:
                 self.back()
             else:
-                self.clear_backup_code()
+                self.clear_credentials()
                 self.set_result(False)
         elif error is Error.FILE_READ_ERROR:
-            self.clear_backup_code()
+            self.clear_credentials()
             await ErrorPage(text='Unable to read backup file.').show()
             self.set_result(False)
         elif error is Error.INVALID_BACKUP_FILE_HEADER:
-            self.clear_backup_code()
+            self.clear_credentials()
             await ErrorPage(text='Unable to read backup file header. The backup may have been modified.').show()
             self.set_result(False)
         elif error is Error.OUT_OF_MEMORY_ERROR:
-            self.clear_backup_code()
+            self.clear_credentials()
             await ErrorPage(text='Not enough memory to verify this backup.').show()
             self.set_result(False)
         else:
-            self.clear_backup_code()
+            self.clear_credentials()
             await ErrorPage(text='Unable to verify backup.').show()
             self.set_result(False)
 
-    def clear_backup_code(self):
+    def clear_credentials(self):
         self.backup_code = [None] * TOTAL_BACKUP_CODE_DIGITS
+        self.backup_password_words = []
+        self.backup_password_prefixes = []
         self.decryption_password = None